+++ src/Vector2D.hpp
@@ -4,11 +4,11 @@
 
 class Vector2D {
 public:
-	uint_fast32_t x;
-	uint_fast32_t y;
+	float x;
+	float y;
 
 	Vector2D();
-	Vector2D(uint_fast32_t x, uint_fast32_t y);
+	Vector2D(float x, float y);
 
 	Vector2D& Add(const Vector2D& vec);
 	Vector2D& Subtract(const Vector2D& vec);
